update sale order services

+++ Backend/2Sport_BE/Controllers/RentalOrderController.cs
@@ -119,7 +119,10 @@ namespace _2Sport_BE.Controllers
             }
             foreach (var item in rentalOrderCM.ProductInformations)
             {
+                if(item.CartItemId != null || item.CartItemId != 0)
+                {
                     await _cartItemService.DeleteCartItem(item.CartItemId);
+                }
             }
             return Ok(response);
         }
